Reformatting, changes to flashcard display width, new flashcards (#51)
* Updated 2016S_FE_AM_5.md

Reformatted the inline math blocks to make the display more appealing.

Reformatted the file (including removing extra spaces and text)

* Updated 2021A_FE_PM_7.1, 2021A_FE_PM_7.2.md

Corrected the position of the ? symbol to correctly split the question and the answer in 7.1 and to make 7.2 recognised as a flashcard.

Fixed the header name to be marked with A instead of S beside the year.

Removed extra space and text.

* Updated data.json

Changed the flashcard display width (flashcardWidthPercentage setting) from 95 to 50. This should prevent images from becoming too large especially on large displays such as desktop devices.

* Updated 2021A_FE_PM_7.1.md

Moved the flashcard separator to a new line to make the separator work in the plugin.

* Added 2025S_FE_AM_2.md

* New flashcards (and a drawing)